Strip to minimal codec-only changeset for upstream review
Remove 15 files that are internal workflow artifacts, not core codec:
- docs: calibration inventory, production audit, pipeline redesign,
  quality analysis, noise model design (5 planning/analysis docs)
- tools: calibration guide, gopro-calibrate script, phocus sequences
- apps: calibrate, fpn_extract, noise_analysis (3 standalone tools)
- scripts: batch_encode, gpr_batch, test_suite, test_noise_compression

What remains (20 new files):
- Core codec: ANS coder, noise model, denoise, encoder/decoder changes
- Validation: ans_test, fuzz_ans, compare_quality
- Docs: architecture, format spec, gotchas, future ideas, compression results
- Test data: gitignore + README for test set organization

+++ source/app/compare_quality/compare_quality.c
@@ -186,7 +186,7 @@ int main(int argc, char **argv) {
     printf("\nError Distribution:\n");
     printf("  Mean error:     %.4f DN (bias)\n", (double)sum_err / n);
     printf("  RMSE:           %.2f DN\n", sqrt(sse / n));
-    printf("  Max error:      %lld DN\n", max_err);
+    printf("  Max error:      %ld DN\n", (long)max_err);
     printf("  Changed pixels: %.1f%%\n", 100.0 * diff_count / n);
     printf("  Within ±1 DN:   %.1f%%\n", 100.0 * (err_hist[0] + err_hist[1]) / n);
     printf("  Within ±5 DN:   %.1f%%\n",
